adjusted index creation and managemen
